Make RemoteObject.subtype more lenient because the protocol lies
The RemoteObject.subtype is not statically known, so the protocol's
enum values are insufficient.
One value that exists for sure and is not listed is 'trustedtype'.
In general, we should make this enum more lenient.

This commit introduces the concept of non-exhaustive enums to mark
that some enums are actually just strings with some known values.

Fixes: https://github.com/joffrey-bion/chrome-devtools-kotlin/issues/546